+++ src/lib/Components/Bidding/Screens/__tests__/ConfirmBid-tests.tsx
@@ -10,6 +10,7 @@ import { Serif14, Serif16 } from "../../Elements/Typography"
 import { BidInfoRow } from "../../Components/BidInfoRow"
 import { Button } from "../../Components/Button"
 import { Checkbox } from "../../Components/Checkbox"
+import { LinkText } from "../../Components/LinkText"
 
 import { BidResultScreen } from "../BidResult"
 import { BillingAddress } from "../BillingAddress"
@@ -148,6 +149,8 @@ describe("when pressing bid button", () => {
     const yourMaxBidRow = component.root.findAllByType(TouchableWithoutFeedback)[0]
     const creditCardRow = component.root.findAllByType(TouchableWithoutFeedback)[1]
     const billingAddressRow = component.root.findAllByType(TouchableWithoutFeedback)[2]
+    const conditionsOfSaleLink = component.root.findByType(LinkText)
+    const conditionsOfSaleCheckbox = component.root.findByType(Checkbox)
 
     yourMaxBidRow.instance.props.onPress()
 
@@ -160,6 +163,8 @@ describe("when pressing bid button", () => {
     billingAddressRow.instance.props.onPress()
 
     expect(navigator.push).not.toHaveBeenCalled()
+    expect(conditionsOfSaleLink.instance.props.onPress).toBeNull()
+    expect(conditionsOfSaleCheckbox.instance.props.disabled).toBeTruthy()
   })
 
   describe("when pressing bid", () => {